requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Jinja2==3.1.6
itsdangerous==2.2.0
blinker==1.9.0
click==8.1.7

opencv-python==4.8.1.78
face_recognition==1.3.0
dlib==19.24.2
face-recognition-models==0.3.0
numpy==1.26.4
Pillow==10.0.1

# # Optional for deployment
# gunicorn==21.2.0
